Add new function to gather all scales in one table in order to use random forest algorithms (at RSU or building scale) Add the RF algorithm in the geoindicator function for LCZ calculation

+++ processingchain/src/main/groovy/org/orbisgis/orbisprocess/geoclimate/processingchain/GeoIndicatorsChain.groovy
@@ -988,7 +988,8 @@ IProcess createUnitsOfAnalysis() {
 /**
  * Compute all geoindicators at the 3 scales :
  * building, block and RSU
- * Compute also the LCZ classification and the urban typology
+ * Compute also the LCZ classification (using the min distance algorithm - lczRandomForest=false -
+ * or the randomForest one - lczRandomForest=true) and the urban typology
  *
  * @return 4 tables outputTableBuildingIndicators, outputTableBlockIndicators, outputTableRsuIndicators,
  * outputTableRsuLcz . The first three tables contains the geoindicators and the last table the LCZ classification.
@@ -1005,11 +1006,13 @@ IProcess computeAllGeoIndicators() {
                 distance: 0.01, indicatorUse: ["LCZ", "URBAN_TYPOLOGY", "TEB"], svfSimplified: false, prefixName: "",
                 mapOfWeights: ["sky_view_factor"             : 1, "aspect_ratio": 1, "building_surface_fraction": 1,
                                "impervious_surface_fraction" : 1, "pervious_surface_fraction": 1,
-                               "height_of_roughness_elements": 1, "terrain_roughness_length": 1]
+                               "height_of_roughness_elements": 1, "terrain_roughness_length": 1],
+                lczRandomForest: false, lczRfModelPath: "../models/LCZ_OSM_RF_1_0.model"
         outputs outputTableBuildingIndicators: String, outputTableBlockIndicators: String,
                 outputTableRsuIndicators: String, outputTableRsuLcz: String, outputTableZone: String
         run { datasource, zoneTable, buildingTable, roadTable, railTable, vegetationTable, hydrographicTable, imperviousTable,
-              surface_vegetation, surface_hydro, distance, indicatorUse, svfSimplified, prefixName, mapOfWeights ->
+              surface_vegetation, surface_hydro, distance, indicatorUse, svfSimplified, prefixName, mapOfWeights, lczRandomForest,
+              lczRfModelPath ->
             info "Start computing the geoindicators..."
             // Temporary tables are created
             def lczIndicTable = postfix "LCZ_INDIC_TABLE"
@@ -1058,7 +1061,8 @@ IProcess computeAllGeoIndicators() {
 
             //Compute block indicators
             def blockIndicators = null
-            if (indicatorUse*.toUpperCase().contains("URBAN_TYPOLOGY")) {
+            if ((indicatorUse*.toUpperCase().contains("URBAN_TYPOLOGY")) ||
+                    (indicatorUse*.toUpperCase().contains("LCZ")) && lczRandomForest) {
                 def computeBlockIndicators = computeBlockIndicators()
                 if (!computeBlockIndicators.execute([datasource            : datasource,
                                                      inputBuildingTableName: buildingIndicators,
@@ -1089,39 +1093,62 @@ IProcess computeAllGeoIndicators() {
 
             // If the LCZ indicators should be calculated, we only affect a LCZ class to each RSU
             if (indicatorUse.contains("LCZ")) {
-                def lczIndicNames = ["GEOM_AVG_HEIGHT_ROOF"              : "HEIGHT_OF_ROUGHNESS_ELEMENTS",
-                                     "BUILDING_FRACTION_LCZ"             : "BUILDING_SURFACE_FRACTION",
-                                     "ASPECT_RATIO"                      : "ASPECT_RATIO",
-                                     "GROUND_SKY_VIEW_FACTOR"            : "SKY_VIEW_FACTOR",
-                                     "PERVIOUS_FRACTION_LCZ"             : "PERVIOUS_SURFACE_FRACTION",
-                                     "IMPERVIOUS_FRACTION_LCZ"           : "IMPERVIOUS_SURFACE_FRACTION",
-                                     "EFFECTIVE_TERRAIN_ROUGHNESS_LENGTH": "TERRAIN_ROUGHNESS_LENGTH"]
+                if (lczRandomForest){
+                    def applygatherScales = Geoindicators.GenericIndicators.gatherScales()
+                    applygatherScales.execute([
+                            buildingTable: buildingIndicators,
+                            blockTable: blockIndicators,
+                            rsuTable: computeRSUIndicators.results.outputTableName,
+                            targetedScale: "RSU",
+                            operationsToApply: ["AVG", "STD"],
+                            prefixName: prefixName,
+                            datasource: datasource])
+                    def gatheredScales = applygatherScales.results.outputTableName
 
-                // Get into a new table the ID, geometry column and the 7 indicators defined by Stewart and Oke (2012)
-                // for LCZ classification (rename the indicators with the real names)
-                def queryReplaceNames = ""
-                lczIndicNames.each { oldIndic, newIndic ->
-                    queryReplaceNames += "ALTER TABLE $lczIndicTable ALTER COLUMN $oldIndic RENAME TO $newIndic;"
+                    def applyRF =  Geoindicators.TypologyClassification.applyRandomForestClassif()
+                    applyRF.execute([
+                            explicativeVariablesTableName   : gatheredScales,
+                            pathAndFileName                 : lczRfModelPath,
+                            idName                          : COLUMN_ID_RSU,
+                            prefixName                      : prefixName,
+                            datasource                      : datasource])
+                    rsuLcz = applyRF.results.outputTableName
                 }
-                datasource.execute """DROP TABLE IF EXISTS $lczIndicTable;
+                else{
+                    def lczIndicNames = ["GEOM_AVG_HEIGHT_ROOF"              : "HEIGHT_OF_ROUGHNESS_ELEMENTS",
+                                         "BUILDING_FRACTION_LCZ"             : "BUILDING_SURFACE_FRACTION",
+                                         "ASPECT_RATIO"                      : "ASPECT_RATIO",
+                                         "GROUND_SKY_VIEW_FACTOR"            : "SKY_VIEW_FACTOR",
+                                         "PERVIOUS_FRACTION_LCZ"             : "PERVIOUS_SURFACE_FRACTION",
+                                         "IMPERVIOUS_FRACTION_LCZ"           : "IMPERVIOUS_SURFACE_FRACTION",
+                                         "EFFECTIVE_TERRAIN_ROUGHNESS_LENGTH": "TERRAIN_ROUGHNESS_LENGTH"]
+
+                    // Get into a new table the ID, geometry column and the 7 indicators defined by Stewart and Oke (2012)
+                    // for LCZ classification (rename the indicators with the real names)
+                    def queryReplaceNames = ""
+                    lczIndicNames.each { oldIndic, newIndic ->
+                        queryReplaceNames += "ALTER TABLE $lczIndicTable ALTER COLUMN $oldIndic RENAME TO $newIndic;"
+                    }
+                    datasource.execute """DROP TABLE IF EXISTS $lczIndicTable;
                                 CREATE TABLE $lczIndicTable 
                                         AS SELECT $COLUMN_ID_RSU, $GEOMETRIC_COLUMN, ${lczIndicNames.keySet().join(",")} 
                                         FROM ${computeRSUIndicators.results.outputTableName};
                                 $queryReplaceNames"""
 
-                // The classification algorithm is called
-                def classifyLCZ = Geoindicators.TypologyClassification.identifyLczType()
-                if (!classifyLCZ([rsuLczIndicators : lczIndicTable,
-                                  rsuAllIndicators : computeRSUIndicators.results.outputTableName,
-                                  normalisationType: "AVG",
-                                  mapOfWeights     : mapOfWeights,
-                                  prefixName       : prefixName,
-                                  datasource       : datasource,
-                                  prefixName       : prefixName])) {
-                    info "Cannot compute the LCZ classification."
-                    return
+                    // The classification algorithm is called
+                    def classifyLCZ = Geoindicators.TypologyClassification.identifyLczType()
+                    if (!classifyLCZ([rsuLczIndicators : lczIndicTable,
+                                      rsuAllIndicators : computeRSUIndicators.results.outputTableName,
+                                      normalisationType: "AVG",
+                                      mapOfWeights     : mapOfWeights,
+                                      prefixName       : prefixName,
+                                      datasource       : datasource,
+                                      prefixName       : prefixName])) {
+                        info "Cannot compute the LCZ classification."
+                        return
+                    }
+                    rsuLcz = classifyLCZ.results.outputTableName
                 }
-                rsuLcz = classifyLCZ.results.outputTableName
             }
 
             datasource.execute "DROP TABLE IF EXISTS $lczIndicTable;"
